pyproject.toml
[project]
name = "gen-ai-projects"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "gradio>=5.9.1",
    "langchain-community>=0.3.13",
    "langchain-openai>=0.2.14",
    "langchain>=0.3.13",
    "langgraph>=0.2.60",
    "omegaconf>=2.3.0",
    "openai>=1.58.1",
    "pandas>=2.2.3",
    "polars>=1.18.0",
    "pyarrow>=18.1.0",
    "pydantic-settings>=2.7.0",
]

[dependency-groups]
dev = [
    "jupyter>=1.1.1",
    "mypy>=1.14.1",
    "nb-black-formatter>=1.0.1",
    "ruff>=0.8.4",
]
